requirements.txt
Django==5.1.3
djangorestframework==3.15.2
pillow==11.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
stripe==11.4.1
forex-python==1.8
redis==5.2.1
celery==5.4.0
django-celery-beat==2.7.0
gunicorn
django-redis
flake8==7.1.1
eventlet==0.39.0
django-cors-headers==4.6.0